Refactor perl critic tooling internals
